pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gworkspace-ai-agent"
version = "0.1.0"
description = "AI Agent for Google Workspace with Human-in-the-Loop approval"
requires-python = ">=3.11"
dependencies = [
    # API Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # LangChain + LangGraph
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",

    # LangChain Google Workspace Tools
    "langchain-google-community[gmail,drive,sheets,calendar]>=2.0.0",
    "langchain-google-genai>=2.0.0",

    # Google APIs (for custom Docs adapter + OAuth)
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",

    # Security
    "cryptography>=42.0",
    "python-jose[cryptography]>=3.3.0",

    # Utilities
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "structlog>=24.0",
    "python-ulid>=2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["backend", "agent", "tools", "hitl"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
